feat: 액티비티 전환 Navigator 이용으로 수정
