+++ feature/createtodolist/build.gradle.kts
@@ -14,8 +14,8 @@ android {
 dependencies {
 
     implementation(libs.androidx.junit.ktx)
-    //implementation(projects.core.database)
-    //implementation(projects.core.model)
+    implementation(projects.core.database)
+    implementation(projects.core.model)
 
     //api(projects.feature.main)
